blog/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-configurations==2.5.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.10
PyJWT==2.10.1
sqlparse==0.5.3
